+++ ck2/cmind/automation.py
@@ -459,7 +459,7 @@ class Automation:
     ############################################################
     def update(self, i):
         """
-        Update Collective Mind artifact
+        Update Collective Mind artifact(s)
 
         Note: not thread safe - we expect one pipeline running on a system
               can make it thread safe when needed (similar to CK)
@@ -486,7 +486,6 @@ class Automation:
 
         # Find CM artifact(s)
         i['out'] = None
-        print (i)
         r = self.search(i)
         if r['return']>0: return r
 
@@ -528,3 +527,113 @@ class Automation:
 
         return {'return':0, 'list':lst}
 
+    ############################################################
+    def move(self, i):
+        """
+        Rename Collective Mind artifacts and/or move them to another CM repository
+
+        Args:
+           parsed_automation
+           parsed_artifact
+
+           parsed_artifacts (cm_obj) 
+
+           meta (dict) - new meta to be merged or to replace the original
+
+        """
+
+        import shutil
+
+        console = i.get('out') == 'con'
+
+        parsed_artifacts = i.get('parsed_artifacts',[])
+
+        if len(parsed_artifacts)==0:
+           return {'return':1, 'error':'target is not specified'}
+        elif len(parsed_artifacts)>1:
+           return {'return':1, 'error':'more than 1 target specified'}
+
+        # Find CM artifact(s)
+        i['out'] = None
+        r = self.search(i)
+        if r['return']>0: return r
+
+        lst = r['list']
+        if len(lst)==0:
+            return {'return':16, 'error':'artifact not found: {}'}
+            
+        # Check target repo
+        target_artifact = parsed_artifacts[0]
+
+        target_artifact_obj = target_artifact[0]
+        target_artifact_repo = target_artifact[1] if len(target_artifact)>0 else None
+
+        r = self.cmind.access({'action':'search',
+                               'automation':'repo',
+                               'artifact': utils.assemble_cm_object2(target_artifact_repo)})
+        if r['return']>0: return r
+
+        target_repo_list = r['list']
+
+        if len(target_repo_list) == 0:
+            return {'return':1, 'error':'target repo "{}" not found'.format(target_artifact_repo)}
+        elif len(target_repo_list) >1:
+            return {'return':1, 'error':'more than 1 target repo found "{}"'.format(target_artifact_repo)}
+
+        target_repo_path = os.path.abspath(target_repo_list[0].path)
+
+        target_artifact_obj_alias = target_artifact_obj[0]
+        target_artifact_obj_uid = target_artifact_obj[1].lower()
+
+
+        # Updating artifacts
+        for artifact in lst:
+
+            artifact_path = os.path.abspath(artifact.path)
+
+            artifact_meta = artifact.original_meta
+
+            artifact_alias = artifact_meta.get('alias','')
+            artifact_uid = artifact_meta.get('uid','')
+
+            artifact_dir_name = os.path.basename(artifact_path)
+            artifact_automation_dir = os.path.dirname(artifact_path)
+            artifact_automation = os.path.basename(artifact_automation_dir)
+
+            must_update_meta = False
+            
+            # Prepare new path
+            new_name = artifact_dir_name
+            if target_artifact_obj_alias != '':
+                new_name = target_artifact_obj_alias
+                
+            new_artifact_path = os.path.join(target_repo_path, artifact_automation, new_name)
+
+            # Check if need to update meta
+            if target_artifact_obj_alias != '' and artifact_alias.lower() != target_artifact_obj_alias.lower():
+                must_update_meta = True
+                artifact_meta['alias']=target_artifact_obj_alias
+
+            if target_artifact_obj_uid != '' and artifact_uid.lower() != target_artifact_obj_uid:
+                must_update_meta = True
+                artifact_meta['uid']=target_artifact_obj_uid
+                
+            artifact.path = new_artifact_path
+
+            # Move
+            if artifact_path != new_artifact_path:
+                if console:
+                    print ('* Moving "{}" to'.format(artifact_path))
+                    print ('         "{}"'.format(new_artifact_path))
+
+                shutil.move(artifact_path, new_artifact_path)
+
+            # Update meta
+            if must_update_meta:
+                if console:
+                    print ('- Updating meta in "{}"'.format(new_artifact_path))
+                
+                r = artifact.update({})
+                if r['return'] >0: return r
+            
+        return {'return':0, 'list':lst}